requirements.txt
torch
torchvision
torchaudio
tqdm
numpy
scikit-learn
matplotlib
pillow
streamlit
